requirements.txt
python-binance
python-dotenv
rich
click
